refactoreo dashboard
