+++ Client/rodarRent/src/App.jsx
@@ -13,8 +13,7 @@ import AboutUs from './views/AboutUs/AboutUs'
 import Admin from './views/Admin/Admin';
 import adminRoutes from './helpers/adminRoutes';
 import Dashboard from './views/Admin/Dashboard';
-import { Clients } from './views/Admin/Clients';
-import {Vehicles} from "./views/Admin/Vehicles"
+
 function App() {
   const [darkMode, setDarkmode] = useState(true)
 
@@ -43,8 +42,6 @@ function App() {
         <Route path={routesHelper.admin} element={<Admin/>}>
          <Route  index element={<Dashboard />} />
           <Route  path={adminRoutes.dashboard} element={<Dashboard />} />
-          <Route  path={adminRoutes.clients} element={<Clients />} />
-          <Route  path={adminRoutes.vehicles} element={<Vehicles />} />
         </Route> 
       </Routes>
     </div>
