--- Client/rodarRent/src/views/Admin/Clients.jsx
@@ -1,3 +1,0 @@
-export const Clients = ()=>{
-    return (<div>Clients</div>)
-}
