--- Client/rodarRent/src/views/Admin/Vehicles.jsx
@@ -1,3 +1,0 @@
-export const Vehicles = ()=>{
-    return (<div>Vehicles</div>)
-}
